added utils for the class
